# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /certificate/create-company-certificate:
    post:
      operationId: CertificateController_generateCompanyCertificate
      parameters: []
      requestBody:
        required: true
        description: Details for creating a company certificate
        content:
          application/json:
            schema:
              type: object
              properties:
                company_ifric_id:
                  type: string
                  example: IFRIC12345
                expiry:
                  type: string
                  format: date
                  example: '2024-12-31'
                user_email:
                  type: string
                  example: user@example.com
              required:
                - company_ifric_id
                - expiry
                - user_email
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Certificate
  /certificate/verify-company-certificate:
    post:
      operationId: CertificateController_verifyCompanyCertificate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_ifric_id:
                  type: string
                  description: Unique identifier for the company
                  example: IFRIC12345
                certificate_data:
                  type: string
                  description: Certificate data to be verified
                  example: base64encodedcertificatestring
              required:
                - company_ifric_id
                - certificate_data
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Certificate
  /certificate/get-company-certificate/{company_ifric_id}:
    get:
      operationId: CertificateController_getCompanyCertificate
      parameters:
        - name: company_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Certificate
  /certificate/reveal-private-key/{company_ifric_id}:
    get:
      operationId: CertificateController_revealPrivateKey
      parameters:
        - name: company_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Certificate
  /certificate/delete-private-key/{company_ifric_id}:
    delete:
      operationId: CertificateController_deletePrivateKey
      parameters:
        - name: company_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Certificate
...
info:
  title: Ifric Registry Service
  description: Open-source multi-tenant company/user/access-control/digital-twin registry service, split into Auth (session/identity), Company (company CRUD, access groups, physical assets), Product (product catalog, digital twins), Certificate, and Script (one-time seed data) controllers. All use a single bearer-JWT auth mechanism (POST /auth/login, refreshed via POST /auth/refresh). Certificate issuance and company_ifric_id minting integrate with an external open-source ICID-compatible service (see README.md).
  version: 0.1.0
  contact: {}
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
...